Add custom web font to replace font-awesome
Font-awesome is a too big so I used fontello to custom add the few icons that I needed, then I
downloaded the zip, copied the code from fontello.css into my styles.scss and added the font folder one
level up to get it working. Now I can add divs with a class of "icon_ICONNAME" where the icon name is given
in fontello.css. Probably not the cleanest way to do it but it seems to work for now.