Tweak social media bar
Decided to let it stay visible for all components, which means I needed to adjust the icon size on smaller
screens and add a margin so that it wouldnt lay over the other components' content.

Also added alt properties to the images in the contact me component and replaced the smaller mountains picture
to test on mobile device (the background looked dull).